finalize a releasable launch method, not perfect
